admin kann res-slots hinzufügen

+++ application/controllers/ajax.php
@@ -16,6 +16,7 @@
             $this->load->model('Courts_model');
 
 
+
             //$this->load->library('pagination');
 
 
@@ -241,6 +242,7 @@
             $id    = $_POST['id'];
 
 
+
             echo $calendar->update($start, $end, $id);
 
         }
@@ -250,13 +252,13 @@
             $calendar = new calendar(DB_HOST, DB_USERNAME, DB_PASSWORD, DATABASE, TABLE);
 
             // Catch start, end and id from javascript
-            $title       = $_POST['title'];
+            $title       =  $_POST['title'];
             $description = $_POST['description'];
             $start_date  = $_POST['start_date'];
             $start_time  = $_POST['start_time'];
             $end_date    = $_POST['end_date'];
             $end_time    = $_POST['end_time'];
-            $color       = $_POST['color'];  //'red'
+            $color       = $_POST['color'];
             $allDay      = $_POST['allDay'];
             $url         = $_POST['url'];
             $user_id      =$_POST['user_id'];
@@ -269,9 +271,14 @@
         public function cal_delete () {
             // Starts the Calendar Class @params 'DB Server', 'DB Username', 'DB Password', 'DB Name', 'Table Name'
             $calendar = new calendar(DB_HOST, DB_USERNAME, DB_PASSWORD, DATABASE, TABLE);
+            $is_admin = $this->ion_auth->user()->row()->permission_group;
+            //$user_id = $this->ion_auth->user()->row()->id;
 
+            if ($is_admin == 'admin') {
             $calendar->delete($_POST['id']);
-
+            }   else {
+                     $calendar->delete_reservation($_POST['id']);
+            }
         }
 
 
@@ -283,7 +290,7 @@
             $id                = $_POST['id'];
             $event_title       = $_POST['title_update'];
             $event_description = $_POST['description_update'];
-            $user_id      = $_POST['user_id'];
+            $active_user_id    = $this->ion_auth->user()->row()->id;
 
             if (isset($_POST['url_update'])) {
                 $url = $_POST['url_update'];
@@ -291,8 +298,7 @@
                 $url = 'false';
             }
 
-            $calendar->updates($id, $event_title, $event_description, $url,  $user_id);
-
+            $calendar->updates($id, $event_title, $event_description, $url,  $active_user_id);
         }
